vot/requirements.txt
vot-trax==3.0.2
tqdm>=4.37
numpy>=1.16
opencv-python>=4.0
six
pylatex>=1.3
jsonschema>=3.2
pyYAML>=5.3
matplotlib>=3.0
Pillow>=7.0
numba>=0.47
requests>=2.22
colorama>=0.4.3
packaging>=20
dominate>=2.5
cachetools>=4.1
bidict>=0.19
phx-class-registry>=3.0
attributee==0.1.3